+++ exercise1/zlasdtest/stack/stack.hpp
@@ -7,7 +7,7 @@
 /* ************************************************************************** */
 
 template <typename Data>
-void Top(uint & testnum, uint & testerr, const lasd::Stack<Data> & stk, bool chk, const Data & val) {
+void Top(unsigned long & testnum, unsigned long & testerr, const lasd::Stack<Data> & stk, bool chk, const Data & val) {
   testnum++;
   bool tst;
   try {
@@ -21,11 +21,11 @@ void Top(uint & testnum, uint & testerr, const lasd::Stack<Data> & stk, bool chk
     tst = false;
     std::cout << std::endl << "Wrong exception: " << exc.what() << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Data>
-void Top(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk, const Data & val) {
+void Top(unsigned long & testnum, unsigned long & testerr, lasd::Stack<Data> & stk, bool chk, const Data & val) {
   testnum++;
   bool tst;
   try {
@@ -39,11 +39,11 @@ void Top(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk, cons
     tst = false;
     std::cout << std::endl << "Wrong exception: " << exc.what() << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Data>
-void Pop(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk) {
+void Pop(unsigned long & testnum, unsigned long & testerr, lasd::Stack<Data> & stk, bool chk) {
   testnum++;
   bool tst = true;
   try {
@@ -58,11 +58,11 @@ void Pop(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk) {
     tst = false;
     std::cout << std::endl << "Wrong exception: " << exc.what() << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Data>
-void TopNPop(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk, const Data & val) {
+void TopNPop(unsigned long & testnum, unsigned long & testerr, lasd::Stack<Data> & stk, bool chk, const Data & val) {
   testnum++;
   bool tst;
   try {
@@ -76,11 +76,11 @@ void TopNPop(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, bool chk, 
     tst = false;
     std::cout << std::endl << "Wrong exception: " << exc.what() << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Data>
-void PushC(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, const Data & val) {
+void PushC(unsigned long & testnum, unsigned long & testerr, lasd::Stack<Data> & stk, const Data & val) {
   testnum++;
   bool tst = true;
   try {
@@ -92,11 +92,11 @@ void PushC(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, const Data &
     std::cout << "\"" << exc.what() << "\": " << "Error!" << std::endl;
     tst = false;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Data>
-void PushM(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, Data val) {
+void PushM(unsigned long & testnum, unsigned long & testerr, lasd::Stack<Data> & stk, Data val) {
   testnum++;
   bool tst = true;
   try {
@@ -108,11 +108,11 @@ void PushM(uint & testnum, uint & testerr, lasd::Stack<Data> & stk, Data val) {
     std::cout << "\"" << exc.what() << "\": " << "Error!" << std::endl;
     tst = false;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Stk>
-void EqualStack(uint & testnum, uint & testerr, const Stk & stk1, const Stk & stk2, bool chk) {
+void EqualStack(unsigned long & testnum, unsigned long & testerr, const Stk & stk1, const Stk & stk2, bool chk) {
   testnum++;
   bool tst;
   try {
@@ -122,11 +122,11 @@ void EqualStack(uint & testnum, uint & testerr, const Stk & stk1, const Stk & st
   catch (std::exception & exc) {
     std::cout << "\"" << exc.what() << "\": " << ((tst = !chk) ? "Correct" : "Error") << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 template <typename Stk>
-void NonEqualStack(uint & testnum, uint & testerr, const Stk & stk1, const Stk & stk2, bool chk) {
+void NonEqualStack(unsigned long & testnum, unsigned long & testerr, const Stk & stk1, const Stk & stk2, bool chk) {
   testnum++;
   bool tst;
   try {
@@ -136,7 +136,7 @@ void NonEqualStack(uint & testnum, uint & testerr, const Stk & stk1, const Stk &
   catch (std::exception & exc) {
     std::cout << "\"" << exc.what() << "\": " << ((tst = !chk) ? "Correct" : "Error") << "!" << std::endl;
   }
-  testerr += (1 - (uint) tst);
+  testerr += (1 - (unsigned long) tst);
 }
 
 /* ************************************************************************** */
